ORC: Cleanup the mess that are the ORC unit tests
- tests failing frequently with key leaks
- clean. it. up.
