+++ h2o-parsers/h2o-orc-parser/src/test/java/water/parser/orc/ParseTestOrc.java
@@ -1,7 +1,6 @@
-package water.parser;
+package water.parser.orc;
 
 import org.junit.Assert;
-import org.junit.Assume;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
@@ -13,18 +12,20 @@ import java.util.TreeSet;
 
 import water.Job;
 import water.Key;
+import water.Scope;
 import water.TestUtil;
 import water.api.schemas3.ParseSetupV3;
 import water.fvec.Frame;
 import water.fvec.NFSFileVec;
 import water.fvec.Vec;
-import water.parser.orc.OrcParserProvider;
+import water.parser.ParseDataset;
+import water.parser.ParseSetup;
+import water.parser.ParseWriter;
 import water.util.ArrayUtils;
-import water.util.FileUtils;
 import water.util.Log;
 
 import static org.junit.Assert.assertEquals;
-import static water.parser.OrcTestUtils.compareOrcAndH2OFrame;
+import static water.parser.orc.OrcTestUtils.compareOrcAndH2OFrame;
 import static water.util.FileUtils.*;
 
 /**
@@ -38,8 +39,8 @@ import static water.util.FileUtils.*;
  */
 public class ParseTestOrc extends TestUtil {
 
-    int totalFilesTested = 0;
-    int numberWrong = 0;
+    private int totalFilesTested = 0;
+    private int numberWrong = 0;
 
     // list all orc files in smalldata/parser/orc directory
     private String[] allOrcFiles = {
@@ -80,70 +81,68 @@ public class ParseTestOrc extends TestUtil {
     @BeforeClass
     static public void setup() { TestUtil.stall_till_cloudsize(1); }
 
-    @BeforeClass
-    static public void _preconditionJavaVersion() { // NOTE: the `_` force execution of this check after setup
-        // Does not run test on Java6 since we are running on Hadoop lib
-        Assume.assumeTrue("Java6 is not supported", !System.getProperty("java.version", "NA").startsWith("1.6"));
-    }
-
-    @Test public void testTypeOverrides(){
-        NFSFileVec nfs = makeNfsFileVec("smalldata/parser/orc/orc_split_elim.orc");
-        Key<Frame> outputKey = Key.make("orc_Test");
-        ParseSetup pstp = new ParseSetup(new ParseSetupV3());
-        pstp._parse_type = new OrcParserProvider.OrcParserInfo();
-        ParseSetup ps = ParseSetup.guessSetup(new Key[]{nfs._key}, pstp);
-        Assert.assertEquals(ps._parse_type.name(), "ORC");
-        ps._column_types[0] = Vec.T_BAD;
-        ps._column_types[1] = Vec.T_CAT;
-        ps._column_types[2] = Vec.T_STR;
-        ParseSetup ps2 = ParseSetup.guessSetup(new Key[]{nfs._key}, ps);
-        String errs = Arrays.deepToString(ps2.errs());
-        Assert.assertEquals(1,ps2.errs().length);
-        Assert.assertTrue(ps2.errs()[0] instanceof ParseWriter.UnsupportedTypeOverride);
-        System.out.println("types: " + Arrays.toString(ArrayUtils.select(Vec.TYPE_STR,ps2._column_types)));
-        Key k = Key.make();
-        Job<Frame> j = ParseDataset.forkParseDataset(k,new Key[]{nfs._key},ps,true)._job;
-        Frame fr = j.get();
-        String  warns = Arrays.toString(j.warns());
-        Assert.assertEquals(errs,warns);
-        Assert.assertArrayEquals(new String[]{"bar", "cat", "dog", "eat", "foo", "zebra"},fr.vec(1).domain());
-        Assert.assertTrue(fr.vec(0).isBad());
-        fr.delete();
+    @Test 
+    public void testTypeOverrides(){
+        Scope.enter();
+        try {
+            NFSFileVec nfs = makeNfsFileVec("smalldata/parser/orc/orc_split_elim.orc");
+            ParseSetup pstp = new ParseSetup(new ParseSetupV3());
+            pstp.setParseType(new OrcParserProvider.OrcParserInfo());
+            ParseSetup ps = ParseSetup.guessSetup(new Key[]{nfs._key}, pstp);
+            Assert.assertEquals(ps.getParseType().name(), "ORC");
+            ps.getColumnTypes()[0] = Vec.T_BAD;
+            ps.getColumnTypes()[1] = Vec.T_CAT;
+            ps.getColumnTypes()[2] = Vec.T_STR;
+            ParseSetup ps2 = ParseSetup.guessSetup(new Key[]{nfs._key}, ps);
+            String errs = Arrays.deepToString(ps2.errs());
+            Assert.assertEquals(1,ps2.errs().length);
+            Assert.assertTrue(ps2.errs()[0] instanceof ParseWriter.UnsupportedTypeOverride);
+            System.out.println("types: " + Arrays.toString(ArrayUtils.select(Vec.TYPE_STR,ps2.getColumnTypes())));
+            Key k = Key.make();
+            Job<Frame> j = ParseDataset.forkParseDataset(k,new Key[]{nfs._key},ps,true)._job;
+            Frame fr = j.get();
+            String  warns = Arrays.toString(j.warns());
+            Assert.assertEquals(errs,warns);
+            Assert.assertArrayEquals(new String[]{"bar", "cat", "dog", "eat", "foo", "zebra"},fr.vec(1).domain());
+            Assert.assertTrue(fr.vec(0).isBad());
+        } finally {
+            Scope.exit();
+        }
     }
 
     @Test
     public void testParseAllOrcs() {
-        Set<String> failedFiles = new TreeSet<>();
-        int numOfOrcFiles = allOrcFiles.length; // number of Orc Files to test
+        try {
+            Set<String> failedFiles = new TreeSet<>();
+            for (String fileName : allOrcFiles) {
+                Log.info("Orc Parser parsing " + fileName);
+                File f = locateFile(fileName);
 
-        for (int fIndex = 0; fIndex < numOfOrcFiles; fIndex++) {
-
-            String fileName = allOrcFiles[fIndex];
-            Log.info("Orc Parser parsing " + fileName);
-            File f = locateFile(fileName);
-
-            if (f != null && f.exists()) {
-                try {
-                    numberWrong += compareOrcAndH2OFrame(fileName, f, failedFiles);
-                    totalFilesTested++;
-                } catch (IOException e) {
-                    e.printStackTrace();
+                if (f != null && f.exists()) {
+                    try {
+                        numberWrong += compareOrcAndH2OFrame(fileName, f, failedFiles);
+                        totalFilesTested++;
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                        failedFiles.add(fileName);
+                        numberWrong++;
+                    }
+                } else {
+                    Log.warn("The following file was not found: " + fileName);
                     failedFiles.add(fileName);
                     numberWrong++;
                 }
-            } else {
-                Log.warn("The following file was not found: " + fileName);
-                failedFiles.add(fileName);
-                numberWrong++;
             }
-        }
 
-        if (numberWrong > 0) {
-            Log.warn("There are errors in your test.");
-            assertEquals("Number of orc files failed to parse is: " + numberWrong + ", failed files = " +
+            if (numberWrong > 0) {
+                Log.warn("There are errors in your test.");
+                assertEquals("Number of orc files failed to parse is: " + numberWrong + ", failed files = " +
                     failedFiles.toString(), 0, numberWrong);
-        } else {
-            Log.info("Parser test passed!  Number of files parsed is " + totalFilesTested);
+            } else {
+                Log.info("Parser test passed!  Number of files parsed is " + totalFilesTested);
+            }
+        } finally {
+            Scope.exit();
         }
     }
 }
